hotfix on the script to add a module, TODO ++

+++ scripts/add_modules.py
@@ -24,8 +24,11 @@ class AddModules(object):
 
         # !!! Always last !!!!
         self.config_db.sadd('modules', module)
-        self.config_db.delete(module + "|" + "parsing")
-        self.config_db.delete(module + "|" + "fetching")
+
+        response = raw_input("Do you want to start a new process for " + module + "? (y/N) ")
+        if response == 'y':
+            self.config_db.delete(module + "|" + "parsing")
+            self.config_db.delete(module + "|" + "fetching")
 
     def from_config_file(self):
         config_file_redis = self.config_file + ".redis"
